Switching to snapshot v2 usage
