+++ Squalr/Source/Prefilters/Prefilter.cs
@@ -47,7 +47,7 @@
         /// <returns>The snapshot generated by the prefilter.</returns>
         public Snapshot GetPrefilteredSnapshot()
         {
-            Snapshot snapshot = SnapshotManager.GetInstance().GetSnapshot(SnapshotRetrievalMode.FromSettings);
+            Snapshot snapshot = SnapshotManagerViewModel.GetInstance().GetSnapshot(SnapshotManagerViewModel.SnapshotRetrievalMode.FromSettings);
 
             foreach (ISnapshotPrefilter prefilter in this.Prefilters)
             {
