+++ Squalr/Source/Snapshots/Snapshot.cs
@@ -1,18 +1,13 @@
 ﻿namespace Squalr.Source.Snapshots
 {
     using SqualrCore.Source.Engine.Types;
-    using SqualrCore.Source.Engine.VirtualMemory;
-    using SqualrCore.Source.Output;
     using SqualrCore.Source.Utils.Extensions;
     using System;
     using System.Collections;
     using System.Collections.Generic;
     using System.Linq;
 
-    /// <summary>
-    /// Defines a snapshot of memory in an external process.
-    /// </summary>
-    internal class Snapshot : IEnumerable
+    internal class Snapshot : IEnumerable<SnapshotRegion>
     {
         /// <summary>
         /// Initializes a new instance of the <see cref="Snapshot" /> class.
@@ -20,9 +15,17 @@
         /// <param name="snapshotName">The snapshot generation method name.</param>
         public Snapshot(String snapshotName = null)
         {
-            this.TimeSinceLastUpdate = DateTime.Now;
             this.SnapshotName = snapshotName == null ? String.Empty : snapshotName;
-            this.SnapshotRegions = new List<SnapshotRegion>();
+            this.MemoryRegions = new List<MemoryRegion>();
+        }
+
+        /// <summary>
+        /// Initializes a new instance of the <see cref="Snapshot" /> class.
+        /// </summary>
+        /// <param name="memoryRegions">The regions with which to initialize this snapshot.</param>
+        public Snapshot(IEnumerable<MemoryRegion> memoryRegions) : this()
+        {
+            this.MemoryRegions = memoryRegions?.ToList();
         }
 
         /// <summary>
@@ -61,11 +64,6 @@
         }
 
         /// <summary>
-        /// Gets the time since the last update was performed on this snapshot.
-        /// </summary>
-        public DateTime TimeSinceLastUpdate { get; private set; }
-
-        /// <summary>
         /// Gets the name associated with the method by which this snapshot was generated.
         /// </summary>
         public String SnapshotName { get; private set; }
@@ -78,7 +76,7 @@
         {
             get
             {
-                return this.SnapshotRegions == null ? 0 : this.SnapshotRegions.Count;
+                return this.SnapshotRegions?.Count() ?? 0;
             }
         }
 
@@ -89,7 +87,7 @@
         {
             get
             {
-                return this.SnapshotRegions == null ? 0L : this.SnapshotRegions.AsEnumerable().Sum(x => x.ByteCount);
+                return this.SnapshotRegions?.Sum(x => x.ByteCount) ?? 0UL;
             }
         }
 
@@ -101,14 +99,56 @@
         {
             get
             {
-                return this.SnapshotRegions == null ? 0 : this.SnapshotRegions.AsEnumerable().Sum(x => x.ElementCount);
+                return this.SnapshotRegions?.Sum(region => region.ElementCount) ?? 0UL;
+            }
+        }
+
+        public DataType LabelDataType
+        {
+            set
+            {
+                this.SnapshotRegions.ForEach(region => region.LabelDataType = value);
+            }
+        }
+
+        public Int32 Alignment
+        {
+            set
+            {
+                this.SnapshotRegions.ForEach(region => region.Alignment = value);
+            }
+        }
+
+        public DataType ElementDataType
+        {
+            set
+            {
+                this.SnapshotRegions.ForEach(region => region.ElementDataType = value);
             }
         }
 
         /// <summary>
-        /// Gets or sets the snapshot regions contained in this snapshot
+        /// Gets or sets the memory regions contained in this snapshot.
         /// </summary>
-        private IList<SnapshotRegion> SnapshotRegions { get; set; }
+        private IList<MemoryRegion> MemoryRegions { get; set; }
+
+        /// <summary>
+        /// Gets the time since the last update was performed on this snapshot.
+        /// </summary>
+        public DateTime TimeSinceLastUpdate { get; private set; }
+
+        private IEnumerable<SnapshotRegion> SnapshotRegions
+        {
+            get
+            {
+                IEnumerator<SnapshotRegion> enumerator = this.GetEnumerator();
+
+                while (enumerator.MoveNext())
+                {
+                    yield return enumerator.Current;
+                }
+            }
+        }
 
         /// <summary>
         /// Indexer to allow the retrieval of the element at the specified index. Notes: This does NOT index into a region.
@@ -133,35 +173,18 @@
                     }
                 }
 
-                OutputViewModel.GetInstance().Log(OutputViewModel.LogLevel.Fatal, "Invalid snapshot index");
                 return null;
             }
         }
 
         /// <summary>
-        /// Sets the data type of the labels contained in this snapshot.
+        /// Creates a shallow clone of this snapshot.
         /// </summary>
-        /// <param name="labelType"></param>
-        public void SetLabelDataType(DataType labelType)
+        /// <param name="newSnapshotName">The snapshot generation method name.</param>
+        /// <returns>The shallow cloned snapshot.</returns>
+        public Snapshot Clone(String newSnapshotName = null)
         {
-            this.SnapshotRegions?.ForEach(x => x.LabelDataType = labelType);
-        }
-
-        /// <summary>
-        /// Updates the alignment and type settings of all snapshot regions.
-        /// </summary>
-        public void UpdateSettings(DataType activeType, Int32 alignment)
-        {
-            if (this.SnapshotRegions == null)
-            {
-                return;
-            }
-
-            foreach (SnapshotRegion region in this)
-            {
-                region.ElementDataType = activeType;
-                region.Alignment = alignment;
-            }
+            return new Snapshot(newSnapshotName, this.SnapshotRegions);
         }
 
         /// <summary>
@@ -180,17 +203,9 @@
         {
             List<SnapshotRegion> validRegions = new List<SnapshotRegion>();
             this.SnapshotRegions?.ForEach(x => validRegions.AddRange(x.GetValidRegions()));
-            this.SnapshotRegions = validRegions;
-        }
 
-        /// <summary>
-        /// Unconditionally expands all regions in this snapshot by the specified size.
-        /// </summary>
-        /// <param name="expandSize">The size by which to expand the snapshot regions.</param>
-        public void ExpandAllRegions(UInt64 expandSize)
-        {
-            this.SnapshotRegions?.ForEach(x => x.Expand(expandSize));
-            this.Intersect(SnapshotManager.GetInstance().GetSnapshot(SnapshotRetrievalMode.FromSettings));
+            // TODO: Something that isn't this
+            // this.SnapshotRegions = validRegions;
         }
 
         /// <summary>
@@ -199,8 +214,8 @@
         public void ReadAllMemory()
         {
             this.TimeSinceLastUpdate = DateTime.Now;
-            this.Intersect(SnapshotManager.GetInstance().GetSnapshot(SnapshotRetrievalMode.FromSettings));
-            this.SnapshotRegions?.ForEach(x => x.ReadAllMemory(keepValues: true));
+            this.Intersect(SnapshotManagerViewModel.GetInstance().GetSnapshot(SnapshotManagerViewModel.SnapshotRetrievalMode.FromSettings));
+            this.SnapshotRegions?.ForEach(x => x.ReadAllMemory());
         }
 
         /// <summary>
@@ -214,33 +229,6 @@
         }
 
         /// <summary>
-        /// Adds snapshot regions to the regions contained in this snapshot. Will automatically merge and sort regions.
-        /// </summary>
-        /// <param name="snapshotRegions">The snapshot regions to add.</param>
-        public void AddSnapshotRegions(IEnumerable<SnapshotRegion> snapshotRegions)
-        {
-            this.AddSnapshotRegions(snapshotRegions?.ToArray());
-        }
-
-        /// <summary>
-        /// Adds snapshot regions to the regions contained in this snapshot. Will automatically merge and sort regions.
-        /// </summary>
-        /// <param name="snapshotRegions">The snapshot regions to add.</param>
-        public void AddSnapshotRegions(params SnapshotRegion[] snapshotRegions)
-        {
-            if (this.SnapshotRegions == null)
-            {
-                this.SnapshotRegions = snapshotRegions.ToList();
-            }
-            else
-            {
-                snapshotRegions?.ForEach(x => this.SnapshotRegions.Add(x));
-            }
-
-            this.MergeAndSortRegions();
-        }
-
-        /// <summary>
         /// Gets the time since the last update was performed on this snapshot.
         /// </summary>
         /// <returns>The time since the last update.</returns>
@@ -250,247 +238,50 @@
         }
 
         /// <summary>
-        /// Gets all snapshot regions contained in this snapshot
-        /// </summary>
-        /// <returns>The snapshot regions contained in this snapshot</returns>
-        public IEnumerable<SnapshotRegion> GetSnapshotRegions()
-        {
-            return this.SnapshotRegions;
-        }
-
-        /// <summary>
-        /// Creates a shallow clone of this snapshot.
-        /// </summary>
-        /// <param name="newSnapshotName">The snapshot generation method name.</param>
-        /// <returns>The shallow cloned snapshot.</returns>
-        public Snapshot Clone(String newSnapshotName = null)
-        {
-            return new Snapshot(newSnapshotName, this.SnapshotRegions);
-        }
-
-        /// <summary>
         /// Determines if an address is contained in this snapshot.
         /// </summary>
         /// <param name="address">The address for which we are searching.</param>
         /// <returns>True if the address is contained.</returns>
         public Boolean ContainsAddress(UInt64 address)
         {
-            if (this.SnapshotRegions == null || this.SnapshotRegions.Count == 0)
-            {
-                return false;
-            }
-
-            return this.ContainsAddressHelper(address, this.SnapshotRegions.Count / 2, 0, this.SnapshotRegions.Count);
-        }
-
-        /// <summary>
-        /// Gets the snapshot region enumerator.
-        /// </summary>
-        /// <returns>The snapshot region enumerator.</returns>
-        public IEnumerator GetEnumerator()
-        {
-            return this.SnapshotRegions.GetEnumerator();
-        }
-
-        /// <summary>
-        /// Helper function for searching for an address in this snapshot. Binary search that assumes this snapshot has sorted regions.
-        /// </summary>
-        /// <param name="address">The address for which we are searching.</param>
-        /// <param name="middle">The middle region index.</param>
-        /// <param name="min">The lower region index.</param>
-        /// <param name="max">The upper region index.</param>
-        /// <returns>True if the address was found.</returns>
-        private Boolean ContainsAddressHelper(UInt64 address, Int32 middle, Int32 min, Int32 max)
-        {
-            if (middle < 0 || middle == this.SnapshotRegions.Count || max < min)
-            {
-                return false;
-            }
-
-            if (address < this.SnapshotRegions.ElementAt(middle).BaseAddress.ToUInt64())
-            {
-                return this.ContainsAddressHelper(address, (min + middle - 1) / 2, min, middle - 1);
-            }
-            else if (address > this.SnapshotRegions.ElementAt(middle).EndAddress.ToUInt64())
-            {
-                return this.ContainsAddressHelper(address, (middle + 1 + max) / 2, middle + 1, max);
-            }
-            else
+            if (this[address] != null)
             {
                 return true;
             }
-        }
 
-        private void Union(Snapshot otherSnapshot)
-        {
-            if (otherSnapshot == null)
-            {
-                return;
-            }
-
-            foreach (SnapshotRegion region in otherSnapshot)
-            {
-                this.SnapshotRegions.Add(region);
-            }
-
-            this.MergeAndSortRegions();
+            return false;
         }
 
         /// <summary>
-        /// Combines the given memory regions with the given memory regions, only keeping the common elements of the two in O(nlogn + n).
+        /// Adds snapshot regions to the regions contained in this snapshot. Will automatically merge and sort regions.
         /// </summary>
-        /// <param name="otherSnapshot">The snapshot regions to mask the target regions against.</param>
-        private void Intersect(Snapshot otherSnapshot)
+        /// <param name="snapshotRegions">The snapshot regions to add.</param>
+        public void AddSnapshotRegions(IEnumerable<SnapshotRegion> snapshotRegions)
         {
-            List<SnapshotRegion> resultRegions = new List<SnapshotRegion>();
-
-            otherSnapshot.MergeAndSortRegions();
-            IEnumerable<NormalizedRegion> otherRegions = otherSnapshot.GetSnapshotRegions();
-
-            if (this.SnapshotRegions.IsNullOrEmpty() || otherRegions.IsNullOrEmpty())
-            {
-                this.SnapshotRegions = resultRegions;
-                return;
-            }
-
-            this.MergeAndSortRegions();
-
-            // Initialize stacks with regions and masking regions
-            Queue<SnapshotRegion> snapshotRegionQueue = new Queue<SnapshotRegion>();
-            Queue<NormalizedRegion> groundTruthQueue = new Queue<NormalizedRegion>();
-
-            // Build candidate region queue from snapshot region array
-            foreach (SnapshotRegion region in this.SnapshotRegions.OrderBy(x => x.BaseAddress.ToUInt64()))
-            {
-                snapshotRegionQueue.Enqueue(region);
-            }
-
-            // Build masking region queue from snapshot
-            foreach (NormalizedRegion maskRegion in otherRegions.OrderBy(x => x.BaseAddress.ToUInt64()))
-            {
-                groundTruthQueue.Enqueue(maskRegion);
-            }
-
-            if (snapshotRegionQueue.Count <= 0 || groundTruthQueue.Count <= 0)
-            {
-                this.SnapshotRegions = resultRegions;
-                return;
-            }
-
-            SnapshotRegion nextSnapshotRegion;
-            NormalizedRegion groundTruthMask = groundTruthQueue.Dequeue();
-
-            while (snapshotRegionQueue.Count > 0)
-            {
-                // Grab next region
-                nextSnapshotRegion = snapshotRegionQueue.Dequeue();
-
-                // Grab the next mask following the current region
-                while (groundTruthMask.EndAddress.ToUInt64() < nextSnapshotRegion.BaseAddress.ToUInt64() && groundTruthQueue.Count > 0)
-                {
-                    groundTruthMask = groundTruthQueue.Dequeue();
-                }
-
-                // Check for mask completely removing this region
-                if (groundTruthMask.EndAddress.ToUInt64() < nextSnapshotRegion.BaseAddress.ToUInt64() || groundTruthMask.BaseAddress.ToUInt64() > nextSnapshotRegion.EndAddress.ToUInt64())
-                {
-                    continue;
-                }
-                // Check for mask completely engulfing this region
-                else if (groundTruthMask.BaseAddress.ToUInt64() <= nextSnapshotRegion.BaseAddress.ToUInt64() && groundTruthMask.EndAddress.ToUInt64() >= nextSnapshotRegion.EndAddress.ToUInt64())
-                {
-                    resultRegions.Add(nextSnapshotRegion);
-                    continue;
-                }
-                // There are no edge cases, we must mask and copy the valid portion of this region
-                else
-                {
-                    UInt64 baseAddress = Math.Max(groundTruthMask.BaseAddress.ToUInt64(), nextSnapshotRegion.BaseAddress.ToUInt64());
-                    UInt64 endAddress = Math.Min(groundTruthMask.EndAddress.ToUInt64(), nextSnapshotRegion.EndAddress.ToUInt64());
-                    Int64 baseOffset = unchecked((Int64)(baseAddress - nextSnapshotRegion.BaseAddress.ToUInt64()));
-
-                    SnapshotRegion newRegion = new SnapshotRegion(nextSnapshotRegion as NormalizedRegion);
-                    newRegion.BaseAddress = baseAddress.ToIntPtr();
-                    newRegion.EndAddress = endAddress.ToIntPtr();
-                    newRegion.SetCurrentValues(nextSnapshotRegion.CurrentValues.LargestSubArray(baseOffset, newRegion.RegionSize.ToInt64()));
-                    newRegion.SetPreviousValues(nextSnapshotRegion.PreviousValues.LargestSubArray(baseOffset, newRegion.RegionSize.ToInt64()));
-                    newRegion.SetElementLabels(nextSnapshotRegion.ElementLabels.LargestSubArray(baseOffset, newRegion.RegionSize.ToInt64()));
-                    newRegion.ElementDataType = nextSnapshotRegion.ElementDataType;
-                    newRegion.Alignment = nextSnapshotRegion.Alignment;
-                    resultRegions.Add(newRegion);
-                }
-            }
-
-            this.SnapshotRegions = resultRegions;
-        }
-
-        private void Except(Snapshot otherSnapshot)
-        {
-            throw new NotImplementedException();
-        }
-
-        private void Subtract(Snapshot otherSnapshot)
-        {
-            throw new NotImplementedException();
+            this.AddSnapshotRegions(snapshotRegions?.ToArray());
         }
 
         /// <summary>
-        /// Merges regions in the current list of memory regions using a fast stack based algorithm O(nlogn + n).
+        /// Adds memory regions to the regions contained in this snapshot. Will automatically merge and sort regions.
         /// </summary>
-        private void MergeAndSortRegions()
+        /// <param name="snapshotRegions">The snapshot regions to add.</param>
+        public void AddSnapshotRegions(params SnapshotRegion[] snapshotRegions)
         {
-            if (this.SnapshotRegions.IsNullOrEmpty())
-            {
-                return;
-            }
+            this.MemoryRegions = new List<MemoryRegion>();
 
-            // First, sort by start address
-            IEnumerable<SnapshotRegion> sortedRegions = this.SnapshotRegions.OrderBy(x => x.BaseAddress.ToUInt64());
+            snapshotRegions?.ForEach(snapshotRegion => this.MemoryRegions.Add(new MemoryRegion(snapshotRegion)));
 
-            // Create and initialize the stack with the first region
-            Stack<SnapshotRegion> combinedRegions = new Stack<SnapshotRegion>();
-            combinedRegions.Push(sortedRegions.First());
+            //// this.MergeAndSortRegions();
+        }
 
-            // Build the remaining regions
-            foreach (SnapshotRegion region in sortedRegions.Skip(1))
-            {
-                SnapshotRegion top = combinedRegions.Peek();
+        public IEnumerator<SnapshotRegion> GetEnumerator()
+        {
+            return this.MemoryRegions?.SelectMany(snapshotRegion => snapshotRegion).GetEnumerator();
+        }
 
-                // If the regions do not overlap, the new region is the top region
-                if (top.EndAddress.ToUInt64() < region.BaseAddress.ToUInt64())
-                {
-                    combinedRegions.Push(region);
-                }
-                // The regions are exactly adjacent; merge them
-                else if (top.EndAddress.ToUInt64() == region.BaseAddress.ToUInt64())
-                {
-                    top.RegionSize = region.EndAddress.Subtract(top.BaseAddress).ToUInt64();
-
-                    // Combine values and labels
-                    top.SetElementLabels(top.ElementLabels?.Concat(region.ElementLabels));
-                    top.SetCurrentValues(top.CurrentValues?.Concat(region.CurrentValues));
-                    top.SetPreviousValues(top.PreviousValues?.Concat(region.PreviousValues));
-                }
-                // The regions overlap
-                else if (top.EndAddress.ToUInt64() <= region.EndAddress.ToUInt64())
-                {
-                    top.RegionSize = region.EndAddress.Subtract(top.BaseAddress).ToUInt64();
-
-                    Int32 overlapSize = unchecked((Int32)(region.EndAddress.ToUInt64() - top.EndAddress.ToUInt64()));
-
-                    // Overlap has conflicting values, so we prioritize the top region and trim the current region
-                    region.SetElementLabels(region.ElementLabels?.SubArray(overlapSize, region.RegionSize.ToInt32() - overlapSize));
-                    region.SetCurrentValues(region.CurrentValues?.SubArray(overlapSize, region.RegionSize.ToInt32() - overlapSize));
-                    region.SetPreviousValues(region.PreviousValues?.SubArray(overlapSize, region.RegionSize.ToInt32() - overlapSize));
-
-                    // Combine values and labels
-                    top.SetElementLabels(top.ElementLabels?.Concat(region.ElementLabels));
-                    top.SetCurrentValues(top.CurrentValues?.Concat(region.CurrentValues));
-                    top.SetPreviousValues(top.PreviousValues?.Concat(region.PreviousValues));
-                }
-            }
-
-            this.SnapshotRegions = combinedRegions.ToList().OrderBy(x => x.BaseAddress.ToUInt64()).ToList();
+        IEnumerator IEnumerable.GetEnumerator()
+        {
+            return this.MemoryRegions?.SelectMany(snapshotRegion => snapshotRegion).GetEnumerator();
         }
     }
     //// End class
