--- Squalr/Source/Snapshots/SnapshotRetrievalMode.cs
@@ -1,15 +1,0 @@
-﻿namespace Squalr.Source.Snapshots
-{
-    using System;
-
-    [Flags]
-    internal enum SnapshotRetrievalMode
-    {
-        FromActiveSnapshot,
-        FromActiveSnapshotOrPrefilter,
-        FromSettings,
-        FromUserModeMemory,
-        FromHeap,
-        FromStack,
-    }
-}
